# backend/contract/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DriveScore API",
    "description": "Usage-based car insurance backend",
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/consent": {
      "post": {
        "tags": [
          "ingestion"
        ],
        "summary": "Give Consent",
        "operationId": "give_consent_v1_consent_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConsentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/trips/start": {
      "post": {
        "tags": [
          "ingestion"
        ],
        "summary": "Start Trip",
        "operationId": "start_trip_v1_trips_start_post",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripStartResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/trips/{trip_id}/chunks": {
      "post": {
        "tags": [
          "ingestion"
        ],
        "summary": "Upload Chunk",
        "operationId": "upload_chunk_v1_trips__trip_id__chunks_post",
        "parameters": [
          {
            "name": "trip_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Trip Id"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TripChunkRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripChunkResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/trips/{trip_id}/end": {
      "post": {
        "tags": [
          "ingestion"
        ],
        "summary": "End Trip",
        "operationId": "end_trip_v1_trips__trip_id__end_post",
        "parameters": [
          {
            "name": "trip_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Trip Id"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripEndResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/trips/{trip_id}/status": {
      "get": {
        "tags": [
          "ingestion"
        ],
        "summary": "Get Trip Status",
        "operationId": "get_trip_status_v1_trips__trip_id__status_get",
        "parameters": [
          {
            "name": "trip_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Trip Id"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/trips/{trip_id}/reprocess": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Reprocess Trip",
        "operationId": "reprocess_trip_v1_trips__trip_id__reprocess_post",
        "parameters": [
          {
            "name": "trip_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Trip Id"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReprocessResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/me": {
      "delete": {
        "tags": [
          "admin"
        ],
        "summary": "Delete My Data",
        "operationId": "delete_my_data_v1_me_delete",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteDriverResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ConsentRequest": {
        "properties": {
          "driver_id": {
            "type": "string",
            "title": "Driver Id"
          },
          "version": {
            "type": "string",
            "title": "Version"
          }
        },
        "type": "object",
        "required": [
          "driver_id",
          "version"
        ],
        "title": "ConsentRequest"
      },
      "ConsentResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "granted_at": {
            "type": "string",
            "format": "date-time",
            "title": "Granted At"
          }
        },
        "type": "object",
        "required": [
          "status",
          "granted_at"
        ],
        "title": "ConsentResponse"
      },
      "DeleteDriverResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "deleted_files": {
            "type": "integer",
            "title": "Deleted Files"
          }
        },
        "type": "object",
        "required": [
          "status",
          "deleted_files"
        ],
        "title": "DeleteDriverResponse"
      },
...
      "GPSSample": {
        "properties": {
          "t": {
            "type": "integer",
            "title": "T"
          },
          "lat": {
            "type": "number",
            "maximum": 90.0,
            "minimum": -90.0,
            "title": "Lat"
          },
          "lon": {
            "type": "number",
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Lon"
          },
          "speed": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Speed"
          },
          "heading": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMaximum": 360.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Heading"
          },
          "accuracy": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Accuracy"
          }
        },
        "type": "object",
        "required": [
          "t",
          "lat",
          "lon"
        ],
        "title": "GPSSample"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "IMUSample": {
        "properties": {
          "t": {
            "type": "integer",
            "title": "T"
          },
          "ax": {
            "type": "number",
            "title": "Ax"
          },
          "ay": {
            "type": "number",
            "title": "Ay"
          },
          "az": {
            "type": "number",
            "title": "Az"
          },
          "gx": {
            "type": "number",
            "title": "Gx"
          },
          "gy": {
            "type": "number",
            "title": "Gy"
          },
          "gz": {
            "type": "number",
            "title": "Gz"
          }
        },
        "type": "object",
        "required": [
          "t",
          "ax",
          "ay",
          "az",
          "gx",
          "gy",
          "gz"
        ],
        "title": "IMUSample"
      },
...
      "ReprocessResponse": {
        "properties": {
          "trip_id": {
            "type": "string",
            "title": "Trip Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "trip_id",
          "status"
        ],
        "title": "ReprocessResponse"
      },
...
      "TripChunkRequest": {
        "properties": {
          "seq": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Seq"
          },
          "imu": {
            "items": {
              "$ref": "#/components/schemas/IMUSample"
            },
            "type": "array",
            "minItems": 1,
            "title": "Imu"
          },
          "gps": {
            "items": {
              "$ref": "#/components/schemas/GPSSample"
            },
            "type": "array",
            "minItems": 1,
            "title": "Gps"
          }
        },
        "type": "object",
        "required": [
          "seq",
          "imu",
          "gps"
        ],
        "title": "TripChunkRequest"
      },
      "TripChunkResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "received_at": {
            "type": "string",
            "format": "date-time",
            "title": "Received At"
          }
        },
        "type": "object",
        "required": [
          "status",
          "received_at"
        ],
        "title": "TripChunkResponse"
      },
...
      "TripEndResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "TripEndResponse"
      },
...
      "TripStartResponse": {
        "properties": {
          "trip_id": {
            "type": "string",
            "title": "Trip Id"
          }
        },
        "type": "object",
        "required": [
          "trip_id"
        ],
        "title": "TripStartResponse"
      },
      "TripStatusResponse": {
        "properties": {
          "trip_id": {
            "type": "string",
            "title": "Trip Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "failure_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Failure Reason"
          }
        },
        "type": "object",
        "required": [
          "trip_id",
          "status"
        ],
        "title": "TripStatusResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}